requirements.txt
requests
playwright
Appium-Python-Client
pytest